requirements.txt
Flask>=0.10.1
requests>=2.5.3
cassandra-driver==2.1.4
blist==1.3.6
